move log envelope creation into log batcher
